Nevermind - returning "false" is working as intended - see description
When the user gets messages that they sent using another person's private key, they will no longer be able to read the message after it has been encrypted. There would have to be another field in the database that stores the message encrypted with the user's own public key to work both ways. Two copies of the message stored overall.